refactor(proto): move js logger to v1
