+++ crates/jstz_proto/src/lib.rs
@@ -2,7 +2,6 @@ mod error;
 
 pub mod context;
 pub mod executor;
-pub mod js_logger;
 pub mod operation;
 pub mod receipt;
 pub mod request_logger;
