+++ crates/jstz_proto/src/runtime/v1/mod.rs
@@ -1,9 +1,11 @@
 mod api;
 mod fetch_handler;
+mod js_logger;
 mod script;
 
 pub use api::{Kv, KvValue, ProtocolApi, ProtocolData, WebApi};
 pub use fetch_handler::{
     fetch, response_from_run_receipt, runtime_and_request_from_run_operation,
 };
+pub use js_logger::{LogRecord, LOG_PREFIX};
 pub use script::ParsedCode;
